scripts/requirements.txt
click
httpx
python-dateutil
beautifulsoup4
